🚚move : update import paths to use new types directory for Troubleshooting and ContactMessage interfaces
